+++ es/api/lazy-dispatch.js
@@ -1,5 +1,0 @@
-import dispatch from '../core/base/dispatch';
-
-export default function (action, payLoadWhenActionIsString, renderKey, delay, option) {
-  dispatch(true, action, payLoadWhenActionIsString, renderKey, delay, option);
-}