+++ src/api/dispatch.js
@@ -1,5 +1,5 @@
 import dispatch from '../core/base/dispatch';
 
-export default function (action, payLoadWhenActionIsString, renderKey, delay, option) {
-  return dispatch(false, action, payLoadWhenActionIsString, renderKey, delay, option);
+export default function (action, payLoadWhenActionIsString, rkOrOptions, delay, extra) {
+  return dispatch(action, payLoadWhenActionIsString, rkOrOptions, delay, extra);
 }