+++ src/api/lazy-reducer.js
@@ -1,5 +1,0 @@
-import ccContext from '../cc-context';
-
-const _lazyReducerCaller = ccContext.reducer._lazyReducerCaller;
-
-export default _lazyReducerCaller;