+++ src/core/base/trigger-computed-and-watch.js
@@ -5,7 +5,7 @@ import watchKeyForRef from '../watch/watch-key-for-ref';
 
 const { store: { getState } } = ccContext;
 
-const callInfo = { payload: null, renderKey: '', delay: -1, noCW: false };
+const callInfo = { payload: null, renderKey: '', delay: -1 };
 
 /** 由首次render触发 */
 export default function (ref) {
@@ -15,18 +15,18 @@ export default function (ref) {
   const connectedModules = okeys(connect);
   const refState = ctx.state;
   if (hasComputedFn) {
-    computeValueForRef(ctx, refModule, refState, refState, callInfo, true);
+    computeValueForRef(ctx, refModule, refState, refState, callInfo, true, true);
     connectedModules.forEach(m => {
       const mState = getState(m);
-      computeValueForRef(ctx, m, mState, mState, callInfo, true);
+      computeValueForRef(ctx, m, mState, mState, callInfo, true, true);
     });
   }
 
   if (hasWatchFn) {
-    watchKeyForRef(ctx, refModule, refState, refState, callInfo, true);
+    watchKeyForRef(ctx, refModule, refState, refState, callInfo, true, true);
     connectedModules.forEach(m => {
       const mState = getState(m);
-      watchKeyForRef(ctx, m, mState, mState, callInfo, true);
+      watchKeyForRef(ctx, m, mState, mState, callInfo, true, true);
     });
   }