+++ src/core/reducer/init-module-reducer.js
@@ -2,7 +2,6 @@ import ccContext from '../../cc-context';
 import * as checker from '../checker';
 import * as util from '../../support/util';
 import dispatch from '../../api/dispatch';
-import lazyDispatch from '../../api/lazy-dispatch';
 import guessDuplicate from '../base/guess-duplicate';
 
 export default function(module, reducer, rootReducerCanNotContainInputModule = true, tag) {
@@ -16,7 +15,7 @@ export default function(module, reducer, rootReducerCanNotContainInputModule = t
   }
 
   const {
-    _reducer, _reducerCaller, _lazyReducerCaller, _reducerFnName_fullFnNames_, _reducerModule_fnNames_,
+    _reducer, _reducerCaller, _reducerFnName_fullFnNames_, _reducerModule_fnNames_,
     // _reducerRefCaller, _lazyReducerRefCaller,
   } = ccContext.reducer;
 
@@ -25,9 +24,7 @@ export default function(module, reducer, rootReducerCanNotContainInputModule = t
   _reducer[module] = newReducer;
   
   const subReducerCaller = util.safeGetObjectFromObject(_reducerCaller, module);
-  const subLazyReducerCaller = util.safeGetObjectFromObject(_lazyReducerCaller, module);
   // const subReducerRefCaller = util.safeGetObjectFromObject(_reducerRefCaller, module);
-  // const subLazyReducerRefCaller = util.safeGetObjectFromObject(_lazyReducerRefCaller, module);
 
   const fnNames = util.safeGetArrayFromObject(_reducerModule_fnNames_, module);
 
@@ -40,15 +37,11 @@ export default function(module, reducer, rootReducerCanNotContainInputModule = t
     if (!fnNames.includes(name)) fnNames.push(name);
     let fullFnName = `${module}/${name}`;
 
-    subReducerCaller[name] = (payload, renderKey, delay) => dispatch(fullFnName, payload, renderKey, delay);
-    subLazyReducerCaller[name] = (payload, renderKey, delay) => lazyDispatch(fullFnName, payload, renderKey, delay);
+    subReducerCaller[name] = (payload, renderKeyOrOptions, delay) => dispatch(fullFnName, payload, renderKeyOrOptions, delay);
 
     // function wrappedReducerFn(payload, delay, idt){
     // }
     // subReducerRefCaller[name] = wrappedReducerFn;
-    // function wrappedLazyReducerFn(payload, delay, idt) {
-    // }
-    // subLazyReducerRefCaller[name] = wrappedLazyReducerFn;
 
     const reducerFn = newReducer[name];
     if (typeof reducerFn !== 'function') {